% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vicoutput_tbl}
\alias{vicoutput_tbl}
\title{Victorian output data}
\format{
A data frame with 30456 rows and 9 variables:
\describe{
\item{\code{sheet}}{The name of Excel sheet the data is extracted from}
\item{\code{output}}{The name of the output}
\item{\code{sub_ouput}}{The disaggregation of the output, if applicable}
\item{\code{measure_type}}{The type or performance measure: being quantity, quality, timeliness or cost.}
\item{\code{unit_of_measure}}{The unit of measure used.}
\item{\code{financial_year}}{The financial year of the data, as a character variable.}
\item{\code{value}}{A numeric being the amount of the measure.}
\item{\code{evalue_type}}{Whether the data is a target, expected or an actual value.}

}
}
\source{
\url{https://www.dtf.vic.gov.au/state-taxation-revenue}
}
\usage{
vicoutput_tbl
}
\description{
Tidy data set of Victorian output performance measures data
}
\keyword{datasets}
