% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{victax_tbl}
\alias{victax_tbl}
\title{Victorian taxation revenue data}
\format{
A data frame with 1564 rows and 8 variables:
\describe{
\item{\code{tax_line}}{The name of the tax line}
\item{\code{tax_sub}}{Sub taxes where tax levies reported separetly}
\item{\code{publication_type}}{The budget, budget update or pre-election budget update publication type}
\item{\code{publication_year}}{The year the budget, budget update or pre-election budget update was for}
\item{\code{financial_year}}{The financial year of the data, as a character variable.}
\item{\code{fy_date}}{The financial year of the data, as a date variable.}
\item{\code{estimate}}{A numeric being the amount of revenue in $ million.}
\item{\code{estimate_type}}{Whether the data is an estimate or an actual value.}

}
}
\source{
\url{https://www.dtf.vic.gov.au/state-taxation-revenue}
}
\usage{
victax_tbl
}
\description{
Tidy data set of Victorian taxation revenue data
}
\keyword{datasets}
